==> step5_affinity_prop.Rmd <==
---
title: "Affinity Propagation Practice"
author: "Noah Klammer"
date: "5/07/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
```

## Clear global env and report

```{r include=FALSE}
rm(list = ls())
gc()
```

# Intro

Here I am experimenting with the package `apcluster`.

# Body

```{r}
library(apcluster)

data("iris")
1:nrow(iris)
alpha <- c(letters,
paste0("a",letters),
paste0("b",letters),
paste0("c",letters),
paste0("d",letters),
paste0("e",letters)
)

# assign row names
rownames(iris) <- alpha[1:nrow(iris)]

apIris1 <- apcluster(negDistMat(r=2), iris)
apIris1
```

```{r}
plot(apIris1, iris)
dev.new(width = 5, height = 10, unit = "in", noRStudioGD = TRUE)
```

```{r}
apcluster::heatmap(apIris1)
```

```{r}
# slot `p` for input preference
# aka q in the parameter argument

# For default q=0.5, this is the median
# of all input simularities
# apIris1@p

# how many exemplars?
length(apIris1@exemplars)

# what is the sixth exemplar?
# ALSO what about labeled variables?
apIris1@exemplars[6]

# how many iterations?
apIris1@it

# what points are in the 2nd cluster?
what <- apIris1@clusters

length(what[[1]])

# how many clusters
length(what)


```


# End

<br><br><br>
